pyproject.toml
[project]
name = "hud-blank"
version = "0.1.0"
description = "A minimal counter-based HUD environment"
requires-python = ">=3.11"
dependencies = [
    "hud-python==0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
